More readable results
It was hard to see where the yellow highlighting of match results began and ended, so I changed the color and added a small margin to the inline results.

In cramped text, the small margin makes the result more clear. I chose green because of green's association with success, and selected a green that keeps the text readable while easily differentiated from the surrounding white background.